+++ src/main/java/com/yuuko/core/commands/interaction/InteractionCommand.java
@@ -1,0 +1,22 @@
+package com.yuuko.core.commands.interaction;
+
+import com.yuuko.core.commands.Command;
+import com.yuuko.core.events.entity.MessageEvent;
+import net.dv8tion.jda.core.Permission;
+
+import java.util.List;
+import java.util.concurrent.ThreadLocalRandom;
+
+public abstract class InteractionCommand extends Command {
+
+    public InteractionCommand(String name, Class<?> module, int expectedParameters, List<String> usage, boolean nsfw, List<Permission> permissions) {
+        super(name, module, expectedParameters, usage, nsfw, permissions);
+    }
+
+    @Override
+    public abstract void onCommand(MessageEvent e);
+
+    public int random(int bound) {
+        return ThreadLocalRandom.current().nextInt(bound);
+    }
+}