+++ src/main/java/com/yuuko/core/commands/interaction/commands/BlushCommand.java
@@ -1,16 +1,15 @@
 package com.yuuko.core.commands.interaction.commands;
 
 import com.yuuko.core.MessageHandler;
-import com.yuuko.core.commands.Command;
+import com.yuuko.core.commands.interaction.InteractionCommand;
 import com.yuuko.core.commands.interaction.InteractionModule;
 import com.yuuko.core.events.entity.MessageEvent;
 import net.dv8tion.jda.core.EmbedBuilder;
 
 import java.util.Arrays;
 import java.util.List;
-import java.util.Random;
 
-public class BlushCommand extends Command {
+public class BlushCommand extends InteractionCommand {
 
     private static final List<String> interactionImage = Arrays.asList(
             "https://i.imgur.com/I3apoUB.gif",
@@ -26,7 +25,7 @@ public class BlushCommand extends Command {
 
     @Override
     public void onCommand(MessageEvent e) {
-        EmbedBuilder embed = new EmbedBuilder().setDescription("**" + e.getMember().getEffectiveName() + "** blushes.").setImage(interactionImage.get(new Random().nextInt(interactionImage.size() -1)));
+        EmbedBuilder embed = new EmbedBuilder().setDescription("**" + e.getMember().getEffectiveName() + "** blushes.").setImage(interactionImage.get(random(interactionImage.size())));
         MessageHandler.sendMessage(e, embed.build());
     }
 }