+++ src/main/java/com/yuuko/core/commands/interaction/commands/DanceCommand.java
@@ -1,7 +1,7 @@
 package com.yuuko.core.commands.interaction.commands;
 
 import com.yuuko.core.MessageHandler;
-import com.yuuko.core.commands.Command;
+import com.yuuko.core.commands.interaction.InteractionCommand;
 import com.yuuko.core.commands.interaction.InteractionModule;
 import com.yuuko.core.events.entity.MessageEvent;
 import com.yuuko.core.utilities.MessageUtilities;
@@ -10,9 +10,8 @@ import net.dv8tion.jda.core.entities.Member;
 
 import java.util.Arrays;
 import java.util.List;
-import java.util.Random;
 
-public class DanceCommand extends Command {
+public class DanceCommand extends InteractionCommand {
 
     private static final List<String> interactionImage = Arrays.asList(
             "https://i.imgur.com/TwMOUGe.gif",
@@ -31,11 +30,11 @@ public class DanceCommand extends Command {
         if(MessageUtilities.checkIfUserMentioned(e)) {
             Member target = MessageUtilities.getMentionedMember(e, true);
             if(target != null) {
-                EmbedBuilder embed = new EmbedBuilder().setDescription("**" + e.getMember().getEffectiveName() + "** dances with **" + target.getEffectiveName() + "**.").setImage(interactionImage.get(new Random().nextInt(interactionImage.size() -1)));
+                EmbedBuilder embed = new EmbedBuilder().setDescription("**" + e.getMember().getEffectiveName() + "** dances with **" + target.getEffectiveName() + "**.").setImage(interactionImage.get(random(interactionImage.size())));
                 MessageHandler.sendMessage(e, embed.build());
             }
         } else {
-            EmbedBuilder embed = new EmbedBuilder().setDescription("**" + e.getMember().getEffectiveName() + "** dances.").setImage(interactionImage.get(new Random().nextInt(interactionImage.size() -1)));
+            EmbedBuilder embed = new EmbedBuilder().setDescription("**" + e.getMember().getEffectiveName() + "** dances.").setImage(interactionImage.get(random(interactionImage.size())));
             MessageHandler.sendMessage(e, embed.build());
         }
     }