+++ src/main/java/com/yuuko/core/commands/interaction/commands/KillCommand.java
@@ -1,7 +1,7 @@
 package com.yuuko.core.commands.interaction.commands;
 
 import com.yuuko.core.MessageHandler;
-import com.yuuko.core.commands.Command;
+import com.yuuko.core.commands.interaction.InteractionCommand;
 import com.yuuko.core.commands.interaction.InteractionModule;
 import com.yuuko.core.events.entity.MessageEvent;
 import com.yuuko.core.utilities.MessageUtilities;
@@ -10,9 +10,8 @@ import net.dv8tion.jda.core.entities.Member;
 
 import java.util.Arrays;
 import java.util.List;
-import java.util.Random;
 
-public class KillCommand extends Command {
+public class KillCommand extends InteractionCommand {
 
     private static final List<String> interactionImage = Arrays.asList(
             "https://i.imgur.com/YGTBd1F.gif",
@@ -30,7 +29,7 @@ public class KillCommand extends Command {
     public void onCommand(MessageEvent e) {
         Member target = MessageUtilities.getMentionedMember(e, true);
         if(target != null) {
-            EmbedBuilder embed = new EmbedBuilder().setDescription("**" + e.getMember().getEffectiveName() + "** kills **" + target.getEffectiveName() + "**.").setImage(interactionImage.get(new Random().nextInt(interactionImage.size() -1)));
+            EmbedBuilder embed = new EmbedBuilder().setDescription("**" + e.getMember().getEffectiveName() + "** kills **" + target.getEffectiveName() + "**.").setImage(interactionImage.get(random(interactionImage.size())));
             MessageHandler.sendMessage(e, embed.build());
         }
     }