+++ CHANGELOG.md
@@ -4,7 +4,8 @@ Key: [+] added, [-] removed, [~] modified, [!] important.
 
 ## 2019_06-11
 + [+] Added [__advice__] command using api.adviceslip.com which is quick and easy.
-
++ [+] Added abstraction to interaction by subclassing Command into InteractionCommand, implementing a LocalThread random function.
+ 
 ## 2019-06-10
 + [~] Updated dependencies to newest possible without breaking other dependencies @JDA && OkHttp
 + [~] Refactored ApplicationProgrammingInterface to simply Api and put it into its own package along with an ApiManager class.